feat(CheckIn): integrate SecurityDashboardActivity
[Delivers #155121723]
